requirements.txt
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
PyMySQL==0.10.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
soupsieve==2.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
yapf==0.30.0
